backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
python-dotenv
websockets
pydantic
httpx
